+++ tools/NewOutputTypeIndex.ps1
@@ -22,7 +22,8 @@ $psd1Files | ForEach {
     Import-LocalizedData -BindingVariable "psd1File" -BaseDirectory $_.DirectoryName -FileName $_.Name
     foreach ($nestedModule in $psd1File.NestedModules)
     {
-        if ($nestedModule.EndsWith("psm1")) {
+        if('.dll' -ne [System.IO.Path]::GetExtension($nestedModule)) 
+        {
             continue;
         }
         $dllPath = Join-Path -Path $_.DirectoryName -ChildPath $nestedModule