+++ src/Synapse/Synapse/Models/Auditing/SqlPoolAuditModel.cs
@@ -1,0 +1,10 @@
+﻿
+namespace Microsoft.Azure.Commands.Synapse.Models
+{
+    public class SqlPoolAuditModel : WorkspaceAuditModel
+    {
+        public string SqlPoolName { get; set; }
+
+        public string[] AuditAction { get; set; }
+    }
+}